implemented Avinash PR review comments
